fix(coreos-init): Finish splitting coreos_startup, properly mount /var
Handle /home and /var with mount units and use systemd's built in tmp
file recreation scheme for creating missing directories in the state
partition. This trims the old script down to just mounting /usr/local in
dev_mode which is fine to leave as-is.

There is a new coreos-startup.target unit which simply depends on every
other unit we want so we don't need add things individually in the
coreos-init ebuild.

Since /var is now backed by /mnt/stateful_partition and tmpfs booting
with a read-only root filesystem might even work now. :)