fix(coreos-init): Disable physical/vm specific units in containers.
When booting a CoreOS filesystem with a container such as systemd-nspawn
there isn't a STATE and OEM partition to mount. All units that cannot or
should not run in this situation are given a "not container" condition
and simply become no-op steps in the boot process. The container is
expected to provide /mnt/stateful_partition since CoreOS still needs it
but cannot mount it itself. Booting looks something like this:

systemd-nspawn --boot --read-only -D ./rootfs \
    --bind ./state:/mnt/stateful_partition