add(coreos-getty-generator): Enable getty on every active console.
Currently systemd (via the systemd-getty-generator) only enables a getty
on a serial terminal if it is mapped to /dev/console. So the kernel
command line "console=tty0 console=ttyS0" will enable a getty on ttyS0
but "console=ttyS0 console=tty0" does not. This behavior surprised me so
here is a generator that will work with both option orders. I also have
a patch to systemd-getty-generator to do the same but unless that patch
is accepted upstream adding a second generator is an easier way to
enable this behavior in CoreOS.