fix(resize_state): Force sector alignment to 1.
gdisk will set the minimum alignment to 8 sectors (4 KB) on all drives
larger than 300 GB just in case it is a newer drive with 4KB physical
sectors. In our case this is terrible because we aren't properly aligned
right now so when resizing for large disks sgdisk shifts the start of
the partition and then everything breaks and puppies cry.