fix(remount): Replace old dev-mode script with two remount units.
The only useful thing the dev-mode script had left in it was remounting
/ and /usr read-write. Hiding those actions in the old 'dev-mode' made
them non-obvious. Also we need to start remounting / read-write all the
time in -usr images all the time, not just dev images. So cleanup time!

On a normal system remounting / and /usr is handled by the
systemd-remount-fs tool but only if those filesystems are in fstab.
Since we don't use fstab I've created two units that will be enabled
depending on what type of image is being generated:

  - prod amd64-generic: neither
  - dev amd64-generic: remount-root
  - prod amd64-usr: remount-root
  - dev amd64-usr: remount-root remount-usr