Cleanup of coreos-install script
Removed bunzip2 --show-progress because this flag is not present on
Ubuntu, Debian, MacPorts, or Centos. As far as I can tell, this flag was
only added in 1.0.6.1 by the MinGW maintainers. This is replaced with
bunzip2 -v.

Changed the fully pathed partprobe to use a nonpathed partprobe because
partprobe lives in /sbin on EL 6.

Changed the message given when the block device is not writable to
remind the user to assume root privileges.