fix(resize-state): Replace bash script with cgpt
Until I sort out the best way to add this to bootengine replace the
current troublesome bash script with the new cgpt implementation.

For now the partition to resize is still selected by filesystem label
but with one behavior difference: The resize is triggered if there is
more than 2MB of space available at the end of the disk rather than an
invalid secondary header. This means that currently valid GPT tables
that do not fill the disk with the STATE partition did not previously
trigger the script but will now.