fix(usr-share-oem): Don't auto-mount OEM on PXE/ISO/etc systems.
The /usr/.noupdate flag is set on systems that don't use our standard
GPT partition layout to block update_engine. On such systems we also
need to block the auto-mounting of OEM since it likely does not exist.

The name 'noupdate' in this context is odd but works for now. I'll
consider changing the logic or flag name when I gut out the old STATE
partition logic.