feat(systemd): Add service to create nsswitch.conf
systemd-tempfiles cannot be used to initialize a completely empty /etc
directory right now because it needs to look up users and groups. Until
a more complete long-term solution is completed on the systemd side of
things for this situation this little one shot will fix this up.

We will need to keep our eye out for similar issues that may be lurking.