udev: properly skip GCE compatibility rule on other systems
Matching GCE instances based on DMI info isn't working properly, a
slightly different syntax is required to match attributes outside of the
current device's tree. I had seen this syntax in use but didn't know
what it meant because it isn't documented. For reference it means:

    "[<SUBSYSTEM>/<KERNEL>]<attribute>"

This slipped through testing because when udev was unable to find the
requested attribute it simply skipped that line rather than interpreting
the value as blank as I had assumed it would. So the rules appeared to
work on GCE and I usually don't run QEMU instances with more than one
network interface.