coreos-install: resolve "current" version
There has been a reported instance of coreos-install downloading the
disk image of 899.15.0 and the signature of 899.17.0, resulting in a
GPG verification failure. This mismatch is due to the fact that our
storage (currently Google Storage) is free to replicate things as slowly
as it would like. In this case, the signature (much smaller) replicated
faster than the disk image.