sshd_keygen: make key generation a bit more robust
Instead of quietly brushing errors under the rug, log at a more verbose
level and exit from the script with an error if there is a problem. The
service unit will now restart the service in an attempt to take a second
stab at key generation.  Additionally, handle the potential case where
no keys are generated (this would have misbehaved with the shell
globbing).