coreos-install: always write a message on failure
Add a few error messages for early failures, and prepend the EXIT
trap with logging of the failed command.  The `getopts` builtin is
writing its own error messages on invalid command arguments, so
that doesn't need any more output.  Also, remove a few messages
that should never be reached.

This fixes coreos/bugs#1441.