coreos-install: If -C is specified without -V, assume "-V current"
Since May 2014 our installation documentation has claimed that
"coreos-install -C <channel>" is the way to install the latest version
in <channel>. However, ever since the -C option was first implemented,
-C has not overridden a defaulted -V. Thus the documented command only
ever worked in the following cases:

1. coreos-install was run on a non-CoreOS host (so we didn't inherit a
   version from os-release).

2. The host was on the desired channel (so -C was redundant and the
   defaulted version existed in that channel). We would succeed but not
   necessarily install the latest version.

3. The host was on a different channel, but the host's version was also
   available in the desired channel due to bit-for-bit promotion. Again
   we wouldn't necessarily get the latest version.

Otherwise it failed on a 404.

In many cases we were probably lucky and users PXE- or ISO-booted an
image that matched the channel they wished to install, but the
documentation does not require this.

Since we no longer ever promote releases without bumping their version,
overriding -C but defaulting -V does not make sense. If we see this
combination, assume "-V current".