*: add ability for test compilation and update .gitignore
Updates the test file to add the ability to compile the tests via
setting the variable `ACTION` to `COMPILE`, this is useful for
running the tests on a Container Linux machine, as you can
generate a compiled test file inside of a docker container with go
and then run them natively.

Updates the .gitignore to add an exception for *.test files.